+++ crates/store-operator/tests/store.rs
@@ -64,7 +64,7 @@ fn store_at(config: SqliteConfig) -> OperatorStore {
 }
 
 fn key(suffix: u8) -> EncodedKey {
-	let mut bytes = 7u64.to_be_bytes().to_vec();
+	let mut bytes = 7u128.to_be_bytes().to_vec();
 	bytes.push(0x10);
 	bytes.push(suffix);
 	EncodedKey::new(bytes)
